modularizacion de verification form
